pyproject.toml
# https://docs.astral.sh/uv/reference/settings/
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/

# PROJECT

[project]
name = "search-agent"
version = "1.0.0"
description = "A Simple Google Search Agent built with Google ADK for Google Cloud."
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a]>=1.10.0",
    "google-cloud-iam>=2.19.1",
    "requests>=2.32.4",
]

# DEPENDENCIES

[dependency-groups]
dev = ["ruff>=0.12.8", "rust-just>=1.42.4"]

# TOOLS

[tool.uv.build-backend]
module-root = ""
module-name = ["search_agent", "user_agent"]

# SYSTEMS

[build-system]
requires = ["uv_build>=0.8.3,<1.0.0"]
build-backend = "uv_build"
